+++ internal/auth/auth.go
@@ -9,18 +9,31 @@ import (
 	"github.com/google/uuid"
 )
 
+// Claims represents the custom claims used in the JWT for authentication.
+// It embeds the standard `jwt.RegisteredClaims` and adds a UserID to identify
+// the subject of the token.
 type Claims struct {
 	jwt.RegisteredClaims
+	// UserID is the unique identifier for the user.
 	UserID string
 }
 
+// AuthConfig holds the configuration and state for JWT-based authentication.
+// It defines the parameters for creating and validating tokens and cookies.
+// It also stores the UserID for the currently authenticated user within a request's context.
 type AuthConfig struct {
+	// CookieName is the name of the HTTP cookie used to store the auth token.
 	CookieName string
-	SecretKey  string
-	TokenExp   time.Duration
-	UserID     string
+	// SecretKey is the secret key used to sign and verify JWTs.
+	SecretKey string
+	// TokenExp is the duration for which a token is valid after being issued.
+	TokenExp time.Duration
+	// UserID holds the identifier of the user making the current request.
+	// It is populated by the authentication middleware/handler after validating a token.
+	UserID string
 }
 
+// NewAuthConfig creates and returns a new AuthConfig instance with default values.
 func NewAuthConfig() AuthConfig {
 	return AuthConfig{
 		CookieName: "auth_token",
@@ -29,29 +42,39 @@ func NewAuthConfig() AuthConfig {
 	}
 }
 
-// BuildJWTString создаёт токен и возвращает его в виде строки.
+// BuildJWTString creates a new JWT for a new user session.
+// It generates a new unique UserID (via UUID), sets the token's expiration time
+// based on `auth.TokenExp`, and signs it using the configured `SecretKey`.
+// It returns the signed token as a string.
 func (auth *AuthConfig) BuildJWTString() (string, error) {
-	// создаём новый токен с алгоритмом подписи HS256 и утверждениями — Claims
 	token := jwt.NewWithClaims(jwt.SigningMethodHS256, Claims{
 		RegisteredClaims: jwt.RegisteredClaims{
-			// когда создан токен
 			ExpiresAt: jwt.NewNumericDate(time.Now().Add(auth.TokenExp)),
 		},
-		// собственное утверждение
 		UserID: uuid.New().String(),
 	})
 
-	// создаём строку токена
 	authToken, err := token.SignedString([]byte(auth.SecretKey))
 	if err != nil {
 		return "", err
 	}
 
-	// возвращаем строку токена
 	return authToken, nil
 }
 
-// FillUserReturnToken возвращает authToken, UserID записывается в auth.UserID
+// FillUserReturnCookie is the central function for handling cookie-based JWT authentication.
+// It inspects the incoming cookie from the request (`incomeCookie`).
+//
+// If the cookie is missing or invalid, it generates a new JWT for a new user session
+// by calling `BuildJWTString`. If the cookie is valid, it parses the JWT to extract
+// the UserID.
+//
+// In both cases, it populates the `auth.UserID` field with the determined user ID, making
+// it available to the rest of the handler chain.
+//
+// It returns a new `http.Cookie` object (with a refreshed expiration) to be set in the response,
+// and an error if one occurred. A special `http.ErrNoCookie` is returned if the
+// token could not be validated.
 func (auth *AuthConfig) FillUserReturnCookie(incomeCookie *http.Cookie) (*http.Cookie, error) {
 	var (
 		resAuthToken string
@@ -64,9 +87,11 @@ func (auth *AuthConfig) FillUserReturnCookie(incomeCookie *http.Cookie) (*http.C
 
 	if len(resAuthToken) == 0 {
 		resAuthToken, err = auth.BuildJWTString()
+		if err != nil {
+			return nil, err
+		}
 	}
 	claims := &Claims{}
-	// парсим из строки токена tokenString в структуру claims
 	jwt.ParseWithClaims(resAuthToken, claims, func(t *jwt.Token) (interface{}, error) {
 		return []byte(auth.SecretKey), nil
 	})
@@ -84,6 +109,5 @@ func (auth *AuthConfig) FillUserReturnCookie(incomeCookie *http.Cookie) (*http.C
 		//Secure:   true,
 		Path: "/",
 	}
-	// возвращаем ID пользователя в читаемом виде
 	return cookie, err
 }
